+++ MOVEit.MIT/Public/Connect-MITServer.ps1
@@ -124,7 +124,11 @@ function Connect-MITServer {
             }
             Write-Output "Connected to MOVEit Transfer server $Hostname"
         }
-    } 
+    }
+    catch [System.Net.Http.HttpRequestException], [System.Net.WebException] {
+        # Format ErrorDetails which contains the JSON response from the REST API
+        $PSCmdlet.ThrowTerminatingError((Format-RestErrorDetails $PSItem))
+    }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)
     }