+++ MOVEit.MIT/Public/Disconnect-MITServer.ps1
@@ -50,6 +50,10 @@ function Disconnect-MITServer {
             "Disconnected from MOVEit Transfer server"
         }
     }
+    catch [System.Net.Http.HttpRequestException], [System.Net.WebException] {
+        # Format ErrorDetails which contains the JSON response from the REST API
+        $PSCmdlet.ThrowTerminatingError((Format-RestErrorDetails $PSItem))
+    }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)
     }